feat: rename cpu module to cpulist
