+++ src/cpulist/parse_test.go
@@ -1,4 +1,4 @@
-package cpu
+package cpulist
 
 import (
 	"reflect"
