+++ src/cpulist/total_test.go
@@ -1,4 +1,4 @@
-package cpu
+package cpulist
 
 import (
 	"testing"
