+++ src/cpulist/validate.go
@@ -1,4 +1,4 @@
-package cpu
+package cpulist
 
 import (
 	"fmt"
