+++ src/pwr_mgmt/pwr_test.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/canonical/rt-conf/src/cpu"
+	"github.com/canonical/rt-conf/src/cpulist"
 	"github.com/canonical/rt-conf/src/data"
 )
 
@@ -101,7 +101,7 @@ func TestPwrMgmt(t *testing.T) {
 
 			for idx, rule := range tc.d {
 
-				parsedCpus, err := cpu.ParseCPUs(rule.CPUs)
+				parsedCpus, err := cpulist.ParseCPUs(rule.CPUs)
 				if err != nil {
 					t.Fatalf("error parsing cpus: %v", err)
 				}
